[composite-bgcolor-animation] EnsureId transparent background
When the background-color is transparent, we would skip the paint
for that. This makes sense for the paint stage because there
really isn't anything to paint. It is all transparent, so we
don't need to paint anything, this optimization makes sense.

Thinking about this more, the actual problem is not missing
the repaint. What we are missing is the UniqueId which is
required to start the animation on the compositor thread.

This CL fixes the problem by ensuring that the UniqueId
exists for the layout object that has background color
animation. In this way, it should fix all the cases where
we might miss a repaint.

Once again, I want to point out that all the optimizations
for repaint makes sense. If we think that we don't need to
repaint (such as the transparent background case), then we
shouldn't. What we are missing is just the UniqueId in order
to start the animation on compositor.

Bug: 1184832
Change-Id: I0a4a675cbe979660de13867defb10b612edbbef1